requirements.txt
customtkinter==5.2.2
faster-whisper==1.1.1
pyannote.audio==3.1.1
torchaudio==2.3.1
torch==2.3.1
huggingface_hub==0.23.5
python-dotenv==1.0.1
tkinterdnd2==0.4.2
numpy==1.26.4
imageio-ffmpeg==0.6.0
